requirements.txt
PySide6
PyInstaller
requests
urllib3
pyserial
geopy
geographiclib
pynmea2